Add precision toggle to ease GPU memory pressure

+++ src/models/whisper.py
@@ -22,10 +22,10 @@ class WhisperModel:
     '''
         Wrapper for Whisper ASR Transcription
     '''
-    def __init__(self, model_name='whisper-small', device=torch.device('cpu'), task='transcribe', language='en'):
+    def __init__(self, model_name='whisper-small', device=torch.device('cpu'), task='transcribe', language='en', dtype=torch.float32):
         # 将自定义名称映射到实际权重名称，并缓存到本地
         self.model_name = model_name
-        self.model = whisper.load_model(MODEL_NAME_MAPPER[model_name], device=device, download_root=CACHE_DIR)
+        self.model = whisper.load_model(MODEL_NAME_MAPPER[model_name], device=device, download_root=CACHE_DIR, dtype=dtype)
         self.task = task                                            # 任务类型：transcribe/translate
         self.language = language.split('_')[0]                      # 取源语言（translate 时为 src）
         self.tokenizer = get_tokenizer(self.model.is_multilingual, num_languages=self.model.num_languages, language=self.language, task=self.task)  # 构造 tokenizer
@@ -48,9 +48,9 @@ class WhisperModelEnsemble:
         Ensemble
         Ensure all models are either multi-lingual or English only
     '''
-    def __init__(self, model_names=['whisper-small'], device=torch.device('cpu'), task='transcribe', language='en'):
+    def __init__(self, model_names=['whisper-small'], device=torch.device('cpu'), task='transcribe', language='en', dtype=torch.float32):
         # 加载多模型并共享 tokenizer，用于集成攻击或评估
-        self.models = [whisper.load_model(MODEL_NAME_MAPPER[model_name], device=device, download_root=CACHE_DIR) for model_name in model_names]  # 逐个载入模型
+        self.models = [whisper.load_model(MODEL_NAME_MAPPER[model_name], device=device, download_root=CACHE_DIR, dtype=dtype) for model_name in model_names]  # 逐个载入模型
         self.task = task                                      # 保存任务类型
         self.language = language                              # 保存语言
         self.tokenizer = get_tokenizer(self.models[0].is_multilingual, num_languages=self.models[0].num_languages, language=self.language, task=self.task)  # 使用首个模型的 tokenizer
